Move patterns to common part;
